--- src/main/java/com/example/anime/domain/character/domain/CharacterState.java
@@ -1,5 +1,0 @@
-package com.example.anime.domain.character.domain;
-
-public enum CharacterState {
-  NOT_MEMORIALIZING, MEMORIALIZING
-}
