+++ src/main/java/com/example/anime/domain/character/dto/request/CharacterRequest.java
@@ -1,0 +1,11 @@
+package com.example.anime.domain.character.dto.request;
+
+public record CharacterRequest (
+        Long animeId,
+        String name,
+        String content,
+        Long lifeTime,
+        String deathReason
+) {
+
+}
