+++ src/main/java/com/example/anime/domain/character/entity/CharacterState.java
@@ -1,0 +1,5 @@
+package com.example.anime.domain.character.entity;
+
+public enum CharacterState {
+  NOT_MEMORIALIZING, MEMORIALIZING
+}
