--- src/main/java/com/example/anime/domain/character/presentation/dto/request/CharacterRequest.java
@@ -1,11 +1,0 @@
-package com.example.anime.domain.character.presentation.dto.request;
-
-public record CharacterRequest (
-        Long animeId,
-        String name,
-        String content,
-        Long lifeTime,
-        String deathReason
-) {
-
-}
